tcp: no error on events
It is clients responsibility to decide what to do with error on onRecv
for example. From tcp layer standpoint; it received data, and raises
event with that data. That event can't be denied. So events handlers
can't return error.

Inside error is reported upstream with onError callback. Client can log
error, use stack trace to show where error occurred. Downstream library
can't know is it appropriate to log, so this is removing logging from
inside library.