udp use single callback function
Instead of many functions for each operation, for success and fail of
the operation. Use single callback with Result type which holds both
operation type, whether it was successful and any success data.

This way we force grouping handling of all io operations in single
function it is not scattered across whole struct. Handling of both
success and failures are at the same place.

Currently I believe that this produces better handlers code.