remove sync cancel example
I added it just to keep track in source control.
I found few problems...
If operation is prepared but not jet submitted we will get NOENT. What
to do in that case; call submit and then cancel again. That is why in
this example is ring.submit called at the top.

If we get ALREADY it is too late for sync cancel then what, we need to
fall back to async and wait for completion.

Sync will work for most of the cases but this edge cases is still not
solved. I can't deallocate structure where operation is stored if there
is still chance that we will need that operation callback.